millestone4/requirements.txt
streamlit==1.25.0
SQLAlchemy==2.0.26
pymysql==1.1.0
python-dotenv==1.0.0
scikit-learn==1.3.1
pydantic==2.5.1
numpy==1.26.0
joblib==1.5.3
shap==0.46.0
lime==0.2.0.1
Pillow==10.0.0
